更新部分服务器代码

+++ document/project/server/returnBook.php
@@ -14,6 +14,7 @@ if(!is_null($_GET['book_isbn']) && !is_null($_GET['user_id']))
 
 function foo()
 {
+	require_once("lock.php");
 	$conn = mysql_open();
 
 	$book_isbn = $_GET['book_isbn'];
@@ -37,7 +38,22 @@ function foo()
 	}
 
 	$result = mysql_fetch_object($query);
-	$book_id = $result -> book_Id;
+	$book_id = $result -> book_id;
+
+	$sql_check = "select borrow_Id from borrowlist where book_id = '$book_id'";
+	$query_check = mysql_query($sql_check);
+	if(!$query_check)
+	{
+		mysql_query("ROLLBACK");
+		$lock -> release();
+		return DATABASE_OPERATION_ERROR;
+	}
+	if(mysql_num_rows($query_check) == 0)
+	{
+		mysql_query("ROLLBACK");
+		$lock -> release();
+		return BOOK_ALL_RETURNED;
+	}
 
 	$sql_return = "delete from borrowlist where book_id = '$book_id' and user_id = '$user_id' limit 1";
 	$query_return = mysql_query($sql_return);
@@ -48,7 +64,7 @@ function foo()
 		return BOOK_ALL_RETURNED;
 	}
 
-	$sql_add_to_stock = "update book set count = count+1 where book_id = 'book_id'";
+	$sql_add_to_stock = "update book_amount set book_amount_available = book_amount_available+1 where book_id = '$book_id'";
 
 	$query_add_to_stock = mysql_query($sql_add_to_stock);
 	if(!$query_add_to_stock)
